Update arguments for `codefresh_api_key` resource
`user_id` is a required parameter to use this resource; considering the diffs between on-prem & SaaS, I didn't feel the need to complicate the examples anymore by adding data call for user, but that is important. 

It should probably be noted that the majority of the [currently documented resources|https://registry.terraform.io/providers/codefresh-io/codefresh/latest/docs/resources/user] do not play well with self-hosted deployments.  I'm testing getting this resource working with a self-hosted deployment and will update as needed if I'm able to get a working model going.