fix: api key read errors out when getting keys for users (#171)
## What
Fix codefresh_api_key resource - failing with error because of an API
behaviour change


## Why

When refreshing state that includes an API key the following error
appeared
<img width="671" height="152" alt="image"
src="https://github.com/user-attachments/assets/ddf726fd-1616-48b7-9d11-13caed9e4a4b"
/>

This is due to the fact the the token request was run without
impersonation. Previously this used to return the proper token, but this
is not the correct behavior. Because tokens are user and account scoped
a switched must be made to the correct user before reading the token.
This PR fixes this on the provider and does the switch the correct user
before reading the token.


## Notes
<!-- Add any notes here -->

## Checklist

* [ ] _I have read
[CONTRIBUTING.md](https://github.com/codefresh-io/terraform-provider-codefresh/blob/master/CONTRIBUTING.md)._
* [ ] _I have [allowed changes to my fork to be
made](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/allowing-changes-to-a-pull-request-branch-created-from-a-fork)._
* [ ] _I have added tests, assuming new tests are warranted_.
* [ ] _I understand that the `/test` comment will be ignored by the CI
trigger [unless it is made by a repo admin or
collaborator](https://codefresh.io/docs/docs/pipelines/triggers/git-triggers/#support-for-building-pull-requests-from-forks)._